refactor: move response module to platform package
• Move response.py from root to platform package structure
• Update client.py imports to use relative import for response
• Fix test imports to reference platform.response module directly
• Rename test_response.py to test_platform_response.py for clarity
• Maintain backward compatibility and functionality
