+++ src/itential_mcp/platform/response.py
